Modified oldgza to work with drop
Previously written packet mangling scripts were changed
to work with drop. Essentially, a game is launched by running
its startgame() function with an integer argument declaring
the VM id it should be forging packets for. To cleanly shut
down, send the process SIGINT and it'll close its nfqueue.
gza.py is now a helper script that takes a gamename and a
VM id and calls the corresponding game with the VM id.